LibJS: Store full realized SourceRange with each AST node
We were spending way too much time converting unrealized source ranges
into line/column pairs on real web content.

This improves JS parsing speed on x.com by 1.13x

+++ Libraries/LibJS/Position.h
@@ -11,9 +11,9 @@
 namespace JS {
 
 struct Position {
-    size_t line { 0 };
-    size_t column { 0 };
-    size_t offset { 0 };
+    u32 line { 0 };
+    u32 column { 0 };
+    u32 offset { 0 };
 };
 
 }
